+++ packages/kit/test/apps/basics/src/routes/routing/trailing-slash-server/+layout.svelte
@@ -3,8 +3,11 @@
 
 	export let data;
 </script>
+
 <ul>
 	<li><a href="/routing/trailing-slash-server/always">/always</a></li>
+	<li><a href="/routing/trailing-slash-server/ignore">/ignore</a></li>
+	<li><a href="/routing/trailing-slash-server/ignore/">/ignore/</a></li>
 	<li><a href="/routing/trailing-slash-server/never/">/never/</a></li>
 </ul>